requirements.txt
scipy==1.12.0
opencv-python==4.9.0.80
keras==2.15.0
tensorflow-cpu==2.15.0
pillow==10.2.0
numpy==1.26.4